backend/requirements.txt
celery==5.2.1
Django==3.2.6
redis==3.5.3
psycopg2-binary==2.8.6
djangorestframework=3.15.2
amqp==5.2.0
argh==0.31.3
asgiref==3.8.1
billiard==3.6.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
drf-yasg==1.21.7
inflection==0.5.1
kombu==5.4.0
packaging==24.1
prompt_toolkit==3.0.47
pytz==2024.1
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
vine==5.1.0
watchdog==2.1.5
wcwidth==0.2.13
